+++ src/Enum/ApiVersion.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Enum;
+
+enum ApiVersion: string
+{
+    case API_V1 = 'API_V1';
+    case API_V2 = 'API_V2';
+}